Add options to configure `labelFilter` and `managedRecordTypes` in Helm Chart
By promoting this options to dedicated values they no longer have to be configured via `extraArgs`

In the [K8GB project](https://github.com/k8gb-io/k8gb), a DNS based load balancer, we use external-dns as a chart dependency.
We would like to configure all values specific to the controller in the default values of our Chart, and leave to the users the provider configuration. This provider configuration usually includes `extraArgs`.
Since `extraArgs` is a list that would be overwritten we would like to keep it empty, otherwise users will have to copy paste our base configuration.