Update aws.md for guidance on latest EKS Auto Mode
AWS EKS Auto Mode recommends an authentication mode of `API` only, which does not leverage the older RBAC config map that this document was referencing. 

This adds a note for any users coming to this page (with the relevant error that is thrown for SEO) so that users have guidance on EKS Auto Mode.