feat(aws): add support for reloadable static creds
When supplying credentials with Hashicorp Vault, users create a file
with static credentials that's referenced by the environment variable
`AWS_SHARED_CREDENTIALS_FILE`. The contents of this file are temporary
credentials that are rotated by Vault on a continuous basis to ensure
they're always valid.

AWS SDK for Go (v2) presently only reads this file once on load, so when
the credentials expire it will never see the updates supplied by Vault.

To fix this, we create a credentials provider which loads the default
config on each retrieval. As a result, the credentials should always be
the latest credentials supplied by Vault.